py script/requirements.txt
# Core PDF/image processing
pymupdf>=1.24.10
pdfplumber>=0.11.4
pillow>=10.3.0
numpy>=1.26.4
opencv-python>=4.10.0.84

# PPTX
python-pptx>=0.6.23

# Data validation
pydantic>=1.10.15

# CLI UX
rich>=13.7.1
tqdm>=4.66.4

# Optional (install later if needed):
# layoutparser
# detectron2
# camelot-py
